Wire through inventory querying
