+++ crates/languages/src/rust.rs
@@ -389,6 +389,7 @@ impl ContextProvider for RustContextProvider {
                     "--".into(),
                     "--nocapture".into(),
                 ],
+                tags: vec!["rust-test".to_owned()],
                 ..TaskTemplate::default()
             },
             TaskTemplate {
